# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StatusResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusResponse",
        "type": "object"
      },
...
      "SubmissionCreate": {
        "properties": {
          "contact_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Id"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "follow_up_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up Date"
          },
          "opportunity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Id"
          },
          "proposed_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Value"
          },
          "subject_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Line"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "SubmissionCreate",
        "type": "object"
      },
      "SubmissionOut": {
        "properties": {
          "contact_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Id"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "follow_up_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up Date"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "opportunity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Id"
          },
          "proposed_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Value"
          },
          "sent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "draft",
            "title": "Status"
          },
          "subject_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Line"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "id"
        ],
        "title": "SubmissionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Canonical API for the Personal OS control plane",
    "title": "Personal Sovereign Enterprise OS",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/submissions": {
      "get": {
        "description": "List submissions using CQRS Query.\n\nThis endpoint demonstrates:\n- Using ListSubmissionsQuery instead of direct DB access\n- Using mediator.send_query() for query execution\n- Using Result type for error handling",
        "operationId": "list_submissions_api_v1_submissions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubmissionOut"
                  },
                  "title": "Response List Submissions Api V1 Submissions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Submissions",
        "tags": [
          "submissions"
        ]
      },
      "post": {
        "description": "Create submission using CQRS Command.\n\nThis endpoint demonstrates:\n- Using CreateSubmissionCommand instead of direct DB insert\n- Using mediator.send_command() for command execution\n- Using Result type for error handling",
        "operationId": "create_submission_api_v1_submissions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Submission",
        "tags": [
          "submissions"
        ]
      }
    },
    "/api/v1/submissions/{sub_id}/mark-lost": {
      "patch": {
        "description": "Mark submission as lost using CQRS Command.\n\nThis endpoint demonstrates:\n- Using MarkSubmissionLostCommand instead of direct DB update\n- Using mediator.send_command() for command execution\n- Using Result type for error handling",
        "operationId": "mark_lost_api_v1_submissions__sub_id__mark_lost_patch",
        "parameters": [
          {
            "in": "path",
            "name": "sub_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sub Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "lost_reason",
            "required": false,
            "schema": {
              "default": "unknown",
              "title": "Lost Reason",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Lost",
        "tags": [
          "submissions"
        ]
      }
    },
    "/api/v1/submissions/{sub_id}/mark-won": {
      "patch": {
        "description": "Mark submission as won using CQRS Command.\n\nThis endpoint demonstrates:\n- Using MarkSubmissionWonCommand instead of direct DB update\n- Using mediator.send_command() for command execution\n- Using Result type for error handling",
        "operationId": "mark_won_api_v1_submissions__sub_id__mark_won_patch",
        "parameters": [
          {
            "in": "path",
            "name": "sub_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sub Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actual_value",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ],
              "default": "0",
              "title": "Actual Value"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Won",
        "tags": [
          "submissions"
        ]
      }
    },
...
  }
}